Rows represented when dragging
Properly recreates the table rows you've selected under the mouse cursor.  Was
toying with the idea of recreating entire table (gaps and all) and picking it
up from the same location you clicked to drag for the helper on draggable, but
it appears getting the cursorAt is a bitch and a half.  Checked in solution
looks nice... good enough.  Time to fucking eat.