Reorganized tables with divs
No longer using tables to layout tables.  Now nested in divs.

Goal is to have headers in a div independent of the tables so we can fix the
height of the div with tables, creating a scrollable fixed-height area.

Keeping the columns in the tables aligned with the headers in the separate div
will be the tricky part, since I set the width of the columns by using headers,
and if I remove the headers from the tables in the scrollable div the column
widths go out of wack.

I can probably manually set them on each column, but it will be a pain.
Perhaps there is a better solution involving shrinking the header so it is
unnoticeable.  Changing visibility to hidden doesn't help, since then its
affects on columns below goes away.