Trying independently scrollable tables
Seems to work.  Need to clean it up.  Might be a better alternative to the
table that scrolls across all rows.

Certainly solves draggable issue encountered when trying to drag upwards and
make it scroll with it.  In this case, you can just scroll any given table to
where you need to drag something and perform the drag + drap.