Just fixing up the groups page
Wanted it to render, so now it just shows the four columns that will ultimately
populate the page.  Gotta work on the back end code to make the request and
filter it properly so we know who goes in what column.

It's entirely too late for me to continue banging on this code.  BEDTIME.