Fixed weird trash issue
I characteristically couldn't put the problem down, and I managed to figure it
out.  Evidently when I wired up the getTrashFlag JS function, I had set up the
back end python to return json.dumps(trash_flag) - which converts the int to
json.  This is clearly not what the back end code wants to read, so all of my
other functions that depended on it returning a python integer got fixed.  Made
a private and exposed version of the function to handle both cases.  Fixed.