Scrolltables Fixed
Draggable shit starting to get implemented. Got rid of callbacks on focus, and
switched to click. Think it should fix some weird issues I was running into -
namely, I require focus to be able to use keyboard events to navigate up and
down within a table.

Seem to be able to drag elements around, and navigate with keyboard very
successfully.