Tiny bug in keyboard movement after sorting
Used synchronizeSelections to handle keeping the table and selections in synch
after a sort. Forgot it requires ids on the rows. Some rows in tables did not
have these - so I've added them throughout. I also added a check for whether or
not there is an id in synchronizeSelections so it winds up being effectively a
no-op, rather than selecting everything in the table like it was.