Breadcrumb patch
Breadcrumbs now work (can build them up and link to them). They take the form:

main,exp_1,sess_1,epoch_1

main being a label for the particular table you're using, the remainders the
individual tables.

Ctrl-click bug solved - when clicking on a given table don't auto-select the
top row. It still selects the table in general, but doesn't highlight a row -
though it enables keyboard movement on the given table. Whether this is the
right-thing-to-do is unclear/arguable, but it does work which may suffice.

Also fixed a bug that I discovered where tables are refreshed twice on the
initial page load (having to do with updating the trash flag after having set
the change callback on the trash radio buttons).

closes #2, closes #6