Popovers now work
Monster commit. Forgive me dear friends, for this may be a beefy read should
you dare to venture in to the diff.

You should find a popover for everything on the browse page. Double click on
each thing, see a form to edit, save and get feedback on it, etc.

Changes are reflected when the popover is closed.

Accomplished with toscawidgets forms. New controllers to handle the
creates/edits. Tried to keep things sane and extensible.