FUSE-based app to expose NIMS as a filesystem.
This first version is working, and has similar functionality to the
shadow fs. However, unlike the shadow fs, the users don't need read
access to all the data to be able to see the data that they have access
to. You can limit the nimsfs mountpoint to be readable only by root and
allow access through nimsfs only. All access limits of NIMS are
implemented, but have not yet been extensively tested, so beware!