Refactored to use fusepy instead of python-fuse
I had some issues getting python-fuse to play nicely with the garbage
collection thread. After some investigation, I got the impression that
the ctypes-based fusepy binding was cleaner and easier to work with than
python-fuse. The codebase for fusepy is certainly simpler-- it's just
one python file (fuse.py). So, I refactored to use it and got the
grabage collection working. So not the caches will not get out of
control. I also added a copy of fuse.py for convenience.