Pfile proc calls prep_convert to find aux files
The pfile pipeline now calls NIMSPfile.prep_convert to request search
criteria that the processor uses to find additional epochs from the same
session that are then passed to convert. This allows nimsraw to request
additiona data files (like calibration scans) that might be needed to
convert the targer pfile.