Update README.mkd
Name of cache_dir is configured as /var/cache/nimsgears in default production.ini (as on scarlet), here /var/cache/nims was created but the absence of /var/cache/nimsgears, with the correct permissions, was preventing the trash contents from being exposed. Perhaps they are both required, but only /var/cache/nimsgears is referenced in any file.