Fixed no locking with 1 job, fix mask issue
No longer lock the DB when num_jobs=1 to help avoid DB lock race
conditions. Also, some non-multiband EPI scans seems to have slices that
are all zeros, causing issues with the code that was intended to clip
away the few bad timepoints that we get with some multiband scans. So,
no longer do that clipping.